Flat assets tree
