+++ app/gui/src/dashboard/layouts/GlobalContextMenu.tsx
@@ -27,7 +27,6 @@ export interface GlobalContextMenuProps {
   readonly category: Category
   readonly rootDirectoryId: DirectoryId
   readonly directoryId: DirectoryId | null
-  readonly path: string | null
   readonly doPaste: (newParentKey: DirectoryId, newParentId: DirectoryId) => void
   readonly event: Pick<React.MouseEvent, 'pageX' | 'pageY'>
 }
@@ -44,7 +43,6 @@ export const GlobalContextMenu = function GlobalContextMenu(props: GlobalContext
     backend,
     category,
     directoryId = null,
-    path,
     rootDirectoryId,
     event,
   } = props
@@ -62,25 +60,25 @@ export const GlobalContextMenu = function GlobalContextMenu(props: GlobalContext
 
   const newFolderRaw = useNewFolder(backend, category)
   const newFolder = useEventCallback(async () => {
-    return await newFolderRaw(directoryId ?? rootDirectoryId, path)
+    return await newFolderRaw(directoryId ?? rootDirectoryId)
   })
-  const newSecretRaw = useNewSecret(backend, category)
+  const newSecretRaw = useNewSecret(backend)
   const newSecret = useEventCallback(async (name: string, value: string) => {
-    return await newSecretRaw(name, value, directoryId ?? rootDirectoryId, path)
+    return await newSecretRaw(name, value, directoryId ?? rootDirectoryId)
   })
   const newProjectRaw = useNewProject(backend, category)
   const newProject = useEventCallback(
     async (templateId: string | null | undefined, templateName: string | null | undefined) => {
-      return await newProjectRaw({ templateName, templateId }, directoryId ?? rootDirectoryId, path)
+      return await newProjectRaw({ templateName, templateId }, directoryId ?? rootDirectoryId)
     },
   )
-  const newDatalinkRaw = useNewDatalink(backend, category)
+  const newDatalinkRaw = useNewDatalink(backend)
   const newDatalink = useEventCallback(async (name: string, value: unknown) => {
-    return await newDatalinkRaw(name, value, directoryId ?? rootDirectoryId, path)
+    return await newDatalinkRaw(name, value, directoryId ?? rootDirectoryId)
   })
   const uploadFilesRaw = useUploadFiles(backend, category)
   const uploadFiles = useEventCallback(async (files: readonly File[]) => {
-    await uploadFilesRaw(files, directoryId ?? rootDirectoryId, path)
+    await uploadFilesRaw(files, directoryId ?? rootDirectoryId)
   })
 
   const entries = (
